+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.2
+
+* Merge pull request [#1](https://github.com/rvndsngwn/map_location_picker/pull/1) from [jeffmilanes/new_update](https://github.com/jeffmilanes)
+* Search results text color changed to black
+* Permission handler removed
+* Code cleanup
+* Readme updated
+
+
 ## 0.0.1+1
 
 * Clean code and performance improvements.